refactor(daemon): move output capture storage into FjallBackend
